+++ tubular/mixins.py
@@ -147,6 +147,7 @@ class WeightColumnMixin:
     def _create_unit_weights_column(
         X: DataFrame,
         return_native: bool = True,
+        verbose: bool = False,
     ) -> tuple[DataFrame, str]:
         """Create unit weights column.
 
@@ -157,9 +158,9 @@ class WeightColumnMixin:
         - does 'unit_weights_column' already exist in data? (unlikely but
         check to be thorough)
         - if it does not, create unit weight 'unit_weights_column'
-        - if it does, is it valid for our purposes? i.e. all unit weights
-        - if it is, then just reuse this existing column
-        - if is not, throw error
+        - if it does, then reuse column
+        - is it valid for our purposes? i.e. all unit weights
+        - if not, raise warning (for verbose=True)
 
         Parameters
         ----------
@@ -169,15 +170,17 @@ class WeightColumnMixin:
         return_native: bool
             controls whether to return nw or pd/pl dataframe
 
-        Raises
-        ------
-            RuntimeError:
-                if invalid 'unit_weights_column' already exists
+        verbose:
+            controls verbosity
 
         Returns
         -------
         DataFrame:
             DataFrame with added 'unit_weights_column'
+
+        Raises
+        ------
+        TypeError: if unit_weights_column already exists and is non numeric.
 
         """
         X = _convert_dataframe_to_narwhals(X)
@@ -185,24 +188,20 @@ class WeightColumnMixin:
         unit_weights_column = "unit_weights_column"
 
         if unit_weights_column in X.columns:
-            all_one = len(X.filter(nw.col(unit_weights_column) == 1)) == len(
-                X,
-            )
-            # if exists already and is valid, return
-            if all_one:
-                return _return_narwhals_or_native_dataframe(
-                    X,
-                    return_native,
-                ), unit_weights_column
+            if X.schema[unit_weights_column] not in NumericTypes:
+                error_msg = f"{unit_weights_column} is present in X and non-numeric, transformer logic requires this to be an all 1 value column."
+                raise TypeError(
+                    error_msg,
+                )
 
-            # error if column already exists but is not suitable
-            msg = "Attempting to insert column of unit weights named 'unit_weights_column', but an existing column shares this name and is not all 1, please rename existing column"
-            raise RuntimeError(
-                msg,
-            )
+            if verbose:
+                # error if column already exists but is not suitable
+                warn_msg = f"column {unit_weights_column} is present in X, transformer logic will assume this column contains all 1 values."
+                warnings.warn(warn_msg, stacklevel=2)
 
-        # finally create dummy weights column if valid option not found
-        X = X.with_columns(nw.lit(1).alias(unit_weights_column))
+        else:
+            # finally create dummy weights column if valid option not found
+            X = X.with_columns(nw.lit(1).alias(unit_weights_column).cast(nw.Int8))
 
         return _return_narwhals_or_native_dataframe(
             X,